% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_username.R
\name{clean_username}
\alias{clean_username}
\title{Clean a user name.}
\usage{
clean_username(username)
}
\arguments{
\item{username}{a complete username}
}
\description{
In practice: keep everything until the first newline
}
