% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posts_texts.R
\name{get_posts_texts}
\alias{get_posts_texts}
\title{Return the main texts of a posts in a website}
\usage{
get_posts_texts(url)
}
\value{
the text of that post
}
\description{
Return the main texts of a posts in a website
}
